virt-v2v: throw away stdout
tried --machine-readable but does not help at all,

since progress is not shown the way the manpage documents on stdout,
but rather (on TW), from the external tool ndbcopy which prints progress on stderr,
and stdout is a formatting mess. Might revisit in the future if we go with SP5.

Signed-off-by: Claudio Fontana <cfontana@suse.de>