testboot_xml.py: testboot_net: reimplement
reimplement testbot_net to detect a successful guest boot to network.

In this reimplementation we look simply for guest network activity
on layer 2. Does not tell us anything else than the fact that
the guest kernel is up and was able to send some packets on layer 2.

Waiting for a better solution to be able to test this at layer 3,
this is what we got for now.

Signed-off-by: Claudio Fontana <cfontana@suse.de>