adjust: fully implement adj_mode "v2v" and "x"
change the behavior to make it fully flexible,
so that conv_mode and adj_mode become almost completely independent.

The only caveat is that when using v2v for the conv_mode,
there is no way to influence the adj_mode.

For all other cases, we can flexibly use v2v for the adj_mode,
or "x", or do no conversion at all.

This is a change in default behavior for previous "x" and "y"
conv_modes, because now we default to "v2v" for the adj_mode,
using "virt-v2v-in-place".

Signed-off-by: Claudio Fontana <cfontana@suse.de>