vmx2xml.py: parse_filename_ref: refuse to use an "auto detect" filename
"auto detect" is used as a placeholder in the .vmx to have the hypervisor
look for the first valid device of a particular type (for example cdrom).

We currently do not have such feature implemented in libvirt, and thus
we do not have such host autodetection feature represented in the xml
either.

Fixes internal gitlab issue #3.

Reported-by: Lin Ma <lma@suse.de>
Signed-off-by: Claudio Fontana <cfontana@suse.de>