virsh_pool_undefine_all.sh: new script to work around virt-install bad behavior
virt-install has a lot of painful problems,
one of which is that it insists in creating storage pools where it should not.

This includes tools like virt-xml, which rely on virt-install.

To work around this, use this script to regularly clean hosts of the
hundreds of useless and harmful pools, which slow down other tools like cockpit.

Signed-off-by: Claudio Fontana <cfontana@suse.de>